requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-multipart==0.0.6
numpy==1.26.4
opencv-python
pillow==9.5.0
tensorflow==2.10.0
segmentation-models>=1.0.1
scikit-learn==1.5.2
patchify
matplotlib==3.9.2
streamlit==1.21.0
requests==2.28.2
pillow==9.5.0
